[compose-navigation] refactor : composable 함수 Graph로 이동
